\name{print.OrderedList}
\alias{print.OrderedList}
\title{ Printing Function for OrderedList Objects }
\description{
The function provides some information about objects that were generated by function \code{\link{OrderedList}}.
}
\usage{
print.OrderedList(x, ...)
}
\arguments{
  \item{x}{ An object of class \code{OrderedList}. }
  \item{...}{ Further printing arguments. }
}
\value{
No value is returned.
}
\references{
Yang X, Bentink S, Scheid S, and Spang R (2006): Similarities of ordered gene lists, to appear in \emph{Journal of Bioinformatics and Computational Biology}.
}
\author{ Stefanie Scheid }
\seealso{ \code{\link{OrderedList}} }
\examples{
data(OL.result)
OL.result
}
\keyword{ print }
